Reorganized login, logout, and session into a new authentication file. Updated tests and views with new authentication file
