+++ supercreative/tests/test_create_course.py
@@ -1,7 +1,7 @@
 from django.core.exceptions import ObjectDoesNotExist
 from django.test import TestCase
 from supercreative.models import (User, Course)
-from supercreative.create_course.create_course import create_course
+from supercreative.create_course.course import create_course
 
 
 class CreateCourseTest(TestCase):
